+++ tests/int_vulnerable_files.rs
@@ -76,7 +76,7 @@ impl TestContext {
                 extraction_depth: 2,
                 no_binary: true,
                 no_extract_archives: false,
-                ignore: Vec::new(),
+                exclude: Vec::new(), // Exclude patterns
             },
             confidence: ConfidenceLevel::Low,
             no_validate: true,
@@ -89,6 +89,8 @@ impl TestContext {
             no_dedup: true,
             ignore_tests: false,
             snippet_length: 256,
+            baseline_file: None,
+            manage_baseline: false,
         };
 
         let loaded = RuleLoader::from_rule_specifiers(&scan_args.rules)
@@ -140,7 +142,7 @@ impl TestContext {
                 extraction_depth: 2,
                 no_binary: true,
                 no_extract_archives: false,
-                ignore: Vec::new(),
+                exclude: Vec::new(), // Exclude patterns
             },
             confidence: ConfidenceLevel::Low,
             no_validate: true,
@@ -153,6 +155,8 @@ impl TestContext {
             no_dedup: true,
             ignore_tests: false,
             snippet_length: 256,
+            baseline_file: None,
+            manage_baseline: false,
         };
 
         let global_args = GlobalArgs {